+++ gradle-plugin/src/main/kotlin/ComposeResourcesPlugin.kt
@@ -1,5 +1,6 @@
 package com.javiersc.compose.resources
 
+import com.javiersc.compose.resources.utils.commonMainKotlin
 import org.gradle.api.Plugin
 import org.gradle.api.Project
 import org.gradle.kotlin.dsl.register
@@ -8,6 +9,8 @@ public class ComposeResourcesPlugin : Plugin<Project> {
 
     override fun apply(project: Project) {
         project.createComposeResourcesExtension().apply { packageName }
+
+        project.commonMainKotlin?.srcDirs(project.generationDir)
 
         project.tasks.register<GenerateComposeResourcesTask>("generateComposeResources")
     }